Send -> SendInput
SendInput is generally the preferred method to send keystrokes and mouse
clicks because of its superior speed and reliability. Under most
conditions, SendInput is nearly instantaneous, even when sending long
strings. Since SendInput is so fast, it is also more reliable because
there is less opportunity for some other window to pop up unexpectedly
and intercept the keystrokes. Reliability is further improved by the
fact that anything the user types during a SendInput is postponed until
afterward.

Unlike the other sending modes, the operating system limits SendInput to
about 5000 characters (this may vary depending on the operating system's
version and performance settings). Characters and events beyond this
limit are not sent.

https://www.autohotkey.com/docs/commands/Send.htm#SendInputDetail